+++ test/tmp/backbone/config/routes.rb
@@ -1,0 +1,3 @@
+Rails.application.routes.draw do
+  mount_i18n_proofreading at: "/i18n_proofreading"
+end